Restructure part 2 - Move components inside pages where they're used (questions and home page) - Create components, containers folders - Fix import paths
