+++ src/app/app.module.ts
@@ -21,22 +21,22 @@ import { MomentModule } from 'angular2-moment'
 import { IonicApp, IonicModule } from 'ionic-angular'
 
 import { MyApp } from './app.component'
-import { AudioInputComponent } from './components/audio-input/audio-input'
-import { CheckboxInputComponent } from './components/checkbox-input/checkbox-input'
-import { InfoScreenComponent } from './components/info-screen/info-screen'
-import { QuestionComponent } from './components/question/question'
-import { RadioInputComponent } from './components/radio-input/radio-input'
-import { RangeInputComponent } from './components/range-input/range-input'
-import { SliderInputComponent } from './components/slider-input/slider-input'
-import { TaskCalendarComponent } from './components/task-calendar/task-calendar'
-import { TaskInfoComponent } from './components/task-info/task-info'
-import { TaskProgressComponent } from './components/task-progress/task-progress'
-import { TickerBarComponent } from './components/ticker-bar/ticker-bar'
-import { TimedTestComponent } from './components/timed-test/timed-test'
 import { ClinicalTasksPage } from './pages/clinical-tasks/clinical-tasks'
 import { EnrolmentPage } from './pages/enrolment/enrolment'
 import { FinishPage } from './pages/finish/finish'
+import { TaskCalendarComponent } from './pages/home/components/task-calendar/task-calendar'
+import { TaskInfoComponent } from './pages/home/components/task-info/task-info'
+import { TaskProgressComponent } from './pages/home/components/task-progress/task-progress'
+import { TickerBarComponent } from './pages/home/components/ticker-bar/ticker-bar'
 import { HomePage } from './pages/home/home'
+import { InfoScreenComponent } from './pages/questions/components/info-screen/info-screen'
+import { AudioInputComponent } from './pages/questions/components/question/audio-input/audio-input'
+import { CheckboxInputComponent } from './pages/questions/components/question/checkbox-input/checkbox-input'
+import { QuestionComponent } from './pages/questions/components/question/question'
+import { RadioInputComponent } from './pages/questions/components/question/radio-input/radio-input'
+import { RangeInputComponent } from './pages/questions/components/question/range-input/range-input'
+import { SliderInputComponent } from './pages/questions/components/question/slider-input/slider-input'
+import { TimedTestComponent } from './pages/questions/components/question/timed-test/timed-test'
 import { QuestionsPage } from './pages/questions/questions'
 import { ReportPage } from './pages/report/report'
 import { SettingsPage } from './pages/settings/settings'
