+++ src/app/pages/home/components/task-info/task-info.scss
