+++ src/app/pages/home/components/task-info/task-info.ts
@@ -14,11 +14,11 @@ import {
   Output
 } from '@angular/core'
 
-import { HomeController } from '../../providers/home-controller'
-import { StorageService } from '../../providers/storage-service'
-import { StorageKeys } from '../../shared/enums/storage'
-import { Task, TasksProgress } from '../../shared/models/task'
-import { checkTaskIsNow } from '../../shared/utilities/check-task-is-now'
+import { HomeController } from '../../../../providers/home-controller'
+import { StorageService } from '../../../../providers/storage-service'
+import { StorageKeys } from '../../../../shared/enums/storage'
+import { Task, TasksProgress } from '../../../../shared/models/task'
+import { checkTaskIsNow } from '../../../../shared/utilities/check-task-is-now'
 
 /**
  * Generated class for the TaskInfo component.
