+++ src/app/pages/home/components/task-progress/task-progress.scss
