+++ src/app/pages/home/components/ticker-bar/ticker-bar.scss
