+++ src/app/pages/questions/components/info-screen/info-screen.scss
