+++ src/app/pages/questions/components/question/audio-input/audio-input.scss
