+++ src/app/pages/questions/components/question/checkbox-input/checkbox-input.scss
