+++ src/app/pages/questions/components/question/question.scss
