+++ src/app/pages/questions/components/question/radio-input/radio-input.html
