+++ src/app/pages/questions/components/question/range-input/range-input.html
