+++ src/app/pages/questions/components/question/slider-input/slider-input.ts
