+++ src/app/pages/questions/components/question/timed-test/timed-test.html
